Remove src/index.ts
The src/index.ts file was removed as it contained javascript code that needed to be translated to TypeScript, but it was bloated and contained types and structures we haven't learned yet. It became clear that refactoring this file was going to be more trouble than starting from scratch, hence the decision to delete it.