Implement new class hierarchy
The old class hierarchy was commented out and a new one was implemented instead in src/index.ts file. The reasons for these changes are the enhancement of encapsulation, the removal of redundant code and the addition of singleton patterns. A new file "homeWork.ts" was created for better organization of code files.