Set Department and Employee class properties to readonly
The properties of the Department and Employee classes have been set as readonly. In the Department class, _name, _domain, and _budget properties are now readonly. Similarly, in the Employee class, _name, _surname, and _paymentInfo properties have been made readonly. This change ensures these properties can only be set during the declaration or within the constructor and prevent unwanted modifications.