Add conditional types in TypeScript
Introduced two new conditional types in a new 'conditional-types/index.ts' file. The first one, 'TReturnType', sets up the type that gets returned from any function. The second type, 'TFuncArgAndReturnType', takes a function type with one or more arguments and returns a tuple where the first value is the function's return type and the second value is the type of its argument.